+++ R/test-files.r
@@ -41,12 +41,16 @@ test_files <- function(paths, reporter = "summary",
 
   current_reporter <- find_reporter(reporter)
   current_reporter$start_reporter()
+
   results <- lapply(paths, test_file, env = env,
     reporter = current_reporter, start_end_reporter = FALSE)
+
   current_reporter$end_reporter()
 
-  results <- unlist(results, recursive = FALSE)
+  # TODO: Integrate results.
+  compiled_results <- test_compiled_code(reporter = reporter)
 
+  results <- unlist(results, recursive = FALSE)
   invisible(testthat_results(results))
 }
 
@@ -179,4 +183,3 @@ sys.source2 <- function(file, envir = parent.frame()) {
 
   invisible(eval(exprs, envir))
 }
-