+++ inst/resources/test-runner.cpp
@@ -1,0 +1,2 @@
+#define TESTTHAT_TEST_RUNNER
+#include <tests/testthat.h>